# docs/swagger.yaml
definitions:
...
  controllers.CreateCommentInput:
    properties:
      comment_text:
        example: this review is rigged!
        type: string
    required:
    - comment_text
    type: object
  controllers.CreateCommentResponse:
    properties:
      comments:
        allOf:
        - $ref: '#/definitions/models.Comment'
        x-order: "1"
      message:
        example: comment created successfully
        type: string
        x-order: "0"
    type: object
  controllers.CreateLikeResponse:
    properties:
      message:
        example: like created successfully
        type: string
        x-order: "0"
    type: object
...
  controllers.DeleteReviewResponse:
    properties:
      message:
        example: review deleted successfully
        type: string
        x-order: "0"
    type: object
...
  controllers.EditReviewInput:
    properties:
      rating:
        example: 1
        type: integer
      review_text:
        example: sample review text
        type: string
    type: object
  controllers.EditReviewResponse:
    properties:
      message:
        example: review updated successfully
        type: string
        x-order: "0"
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
        x-order: "1"
    type: object
...
  controllers.GetReviewLikesResponse:
    properties:
      count:
        example: 1
        type: integer
        x-order: "0"
      rows:
        items:
          $ref: '#/definitions/models.Like'
        type: array
        x-order: "1"
    type: object
...
  models.Comment:
    properties:
      comment_id:
        example: 1
        type: integer
      comment_text:
        example: sample comment text
        type: string
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
      review_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  models.Like:
    properties:
      comment_id:
        example: 1
        type: integer
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
      like_id:
        example: 1
        type: integer
      review_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  models.Review:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
        x-order: "5"
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
        x-order: "5"
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
        x-order: "6"
      phone_id:
        example: 1
        type: integer
        x-order: "1"
      rating:
        example: 8
        type: integer
        x-order: "3"
      review_id:
        example: 1
        type: integer
        x-order: "0"
      review_text:
        example: sample review text
        type: string
        x-order: "4"
      user_id:
        example: 1
        type: integer
        x-order: "2"
    type: object
info:
  contact: {}
paths:
...
  /reviews/{review_id}:
    delete:
      description: Delete a review
      parameters:
      - description: ReviewID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DeleteReviewResponse'
      security:
      - Bearer: []
      summary: Delete a review
      tags:
      - Reviews
    put:
      description: Update a review
      parameters:
      - description: ReviewID
        in: path
        name: review_id
        required: true
        type: string
      - description: the body to edit phone review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.EditReviewInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.EditReviewResponse'
      security:
      - Bearer: []
      summary: Update a review
      tags:
      - Reviews
  /reviews/{review_id}/comments:
    get:
      description: Get all review comment by review ID
      parameters:
      - description: ReviewID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
      summary: Get all review comment by review ID
      tags:
      - Reviews
    post:
      description: Create a review comment by review ID
      parameters:
      - description: ReviewID
        in: path
        name: review_id
        required: true
        type: string
      - description: the body to create phone review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateCommentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CreateCommentResponse'
      security:
      - Bearer: []
      summary: Create a review comment by review ID
      tags:
      - Reviews
  /reviews/{review_id}/likes:
    get:
      description: Get all review likes by review ID
      parameters:
      - description: ReviewID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetReviewLikesResponse'
      summary: Get all review likes by review ID
      tags:
      - Reviews
    post:
      description: Create a review like by review ID
      parameters:
      - description: ReviewID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.CreateLikeResponse'
      security:
      - Bearer: []
      summary: Create a review like by review ID
      tags:
      - Reviews
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token (for example: "Bearer
      eyxyz").'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"